document password hashing
