+++ src/domain/model/register_user.rs
@@ -1,4 +1,4 @@
-use crate::api::utils::password_hasher;
+use crate::api::utils::security;
 
 use super::{
     auth_repo_errors::AuthRepositoryError, user_email::UserEmail, user_password::UserPassword,
@@ -48,7 +48,7 @@ pub struct PasswordHashingError;
 impl HashedUserPassword {
     pub fn new(password: UserPassword) -> Result<HashedUserPassword, PasswordHashingError> {
         let hashed_password =
-            password_hasher::hash_password(password.get()).map_err(|_| PasswordHashingError)?;
+            security::hash_password(password.get()).map_err(|_| PasswordHashingError)?;
         Ok(HashedUserPassword(hashed_password))
     }
